Separate data source and controller functionality
