+++ src/db/data_source.ts
@@ -15,14 +15,3 @@ export const AppDataSource = new DataSource({
         entities: [User, Account, Branch, AccountType, Transaction, TransactionCategory],
         synchronize: true,
     })
-    
-AppDataSource.initialize()
-      .then(async () => {
-          console.log("Data Source has been initialized!")
-          const userRepository = AppDataSource.getRepository(User);
-          const rows = await userRepository.find();
-          console.log(rows)
-      })
-      .catch((err) => {
-          console.error("Error during Data Source initialization", err)
-      })
